+++ src/hypothesis_graphql/strategies.py
@@ -1,7 +1,6 @@
 # pylint: disable=unused-import
-from ._strategies.mutations import mutations
-from ._strategies.queries import queries
 from ._strategies.schema import schemas
+from ._strategies.strategy import mutations, queries
 
 # Backward compatibility
 query = queries